+++ 2D top down/EventHandle.cpp
@@ -498,7 +498,7 @@ void EventHandle(SDL_Event& event)
 
 	testPlayer->MovePlayerAccordingToInput();
 
-	testPlayer->directionFacing = GetAngleAsDegrees(testPlayer->xLoc, testPlayer->yLoc, mouse.x, mouse.y);
+	testPlayer->directionFacing = (float)GetAngleAsDegrees(testPlayer->xLoc, testPlayer->yLoc, mouse.x, mouse.y);
 
 	// Fire weapon if needed.
 	if (mouse.left)
